Add Train Method
